feat: move tex surface stuff to `tex::surface`
